refactor: move guards, pipes into apps folder
