--- apps/bot/src/apps/deal/deal.error-message.ts
@@ -1,3 +1,0 @@
-export enum DealErrorMessage {
-  NOT_FOUND_ATTACHMENTS = 'Not found attachments',
-}
