+++ apps/bot/src/apps/guild/guards/guild.guard.ts
