+++ apps/bot/src/apps/offer/clients/offer.client.ts
@@ -3,7 +3,8 @@ import { CreateOfferInput } from '@lib/domains/offer/application/commands/create
 import { CreateOfferCommand } from '@lib/domains/offer/application/commands/create-offer/create-offer.command';
 import { UpdateOfferInput } from '@lib/domains/offer/application/commands/update-offer/update-offer.input';
 import { UpdateOfferCommand } from '@lib/domains/offer/application/commands/update-offer/update-offer.command';
-import { UserImageClient } from '../user-image/user-image.client';
+import { DeleteOfferCommand } from '@lib/domains/offer/application/commands/delete-offer/delete-offer.command';
+import { UserImageClient } from '../../user-image/clients/user-image.client';
 
 @Injectable()
 export class OfferClient extends UserImageClient {
@@ -13,5 +14,9 @@ export class OfferClient extends UserImageClient {
 
   async updateOffer(input: UpdateOfferInput) {
     await this.commandBus.execute(new UpdateOfferCommand(input));
+  }
+
+  async deleteOffer(id: string) {
+    await this.commandBus.execute(new DeleteOfferCommand(id));
   }
 }
