+++ apps/bot/src/apps/offer/parsers/offer.error-message.ts
