+++ apps/bot/src/apps/offer/parsers/offer.types.ts
