+++ apps/bot/src/apps/offer/pipes/parse-update-offer-input.pipe.ts
@@ -1,7 +1,7 @@
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
 import { UpdateOfferInput } from '@lib/domains/offer/application/commands/update-offer/update-offer.input';
-import { UserWithMessage } from '../../apps/user/user.types';
-import { OfferParser } from '../../apps/offer/offer.parser';
+import { UserWithMessage } from '../../user/parsers/user.types';
+import { OfferParser } from '../parsers/offer.parser';
 
 @Injectable()
 export class ParseUpdateOfferInputPipe implements PipeTransform {
