+++ apps/bot/src/apps/user/parsers/user.types.ts
