+++ apps/bot/src/apps/user/pipes/parse-user-with-message.pipe.ts
@@ -1,11 +1,11 @@
-import { UserClient } from '@app/bot/apps/user/user.client';
+import { UserClient } from '@app/bot/apps/user/clients/user.client';
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
 import { RpcException } from '@nestjs/microservices';
 import { ContextOf } from 'necord';
-import { UserWithMessage } from '@app/bot/apps/user/user.types';
-import { UserErrorMessage } from '@app/bot/apps/user/user.error-message';
-import { UserParser } from '@app/bot/apps/user/user.parser';
-import { Message } from 'discord.js';
+import { UserWithMessage } from '@app/bot/apps/user/parsers/user.types';
+import { UserErrorMessage } from '@app/bot/apps/user/parsers/user.error-message';
+import { UserParser } from '@app/bot/apps/user/parsers/user.parser';
+import { Message, PartialMessage } from 'discord.js';
 
 @Injectable()
 export class ParseUserWithMessagePipe implements PipeTransform {
@@ -18,7 +18,8 @@ export class ParseUserWithMessagePipe implements PipeTransform {
     [createdOrOldmessage, newMessage]: ContextOf<'messageCreate' | 'messageUpdate'>,
     metadata: ArgumentMetadata,
   ): Promise<UserWithMessage> {
-    let message: Message;
+    let message: Message | PartialMessage;
+
     if (newMessage) message = await newMessage.fetch();
     else message = await createdOrOldmessage.fetch();
