+++ apps/bot/src/apps/user/pipes/parse-user.pipe.ts
@@ -1,8 +1,8 @@
-import { UserClient } from '@app/bot/apps/user/user.client';
+import { UserClient } from '@app/bot/apps/user/clients/user.client';
 import { ArgumentMetadata, Injectable, PipeTransform } from '@nestjs/common';
 import { ContextOf } from 'necord';
-import { UserParser } from '@app/bot/apps/user/user.parser';
-import { SimpleUser } from '@app/bot/apps/user/user.types';
+import { UserParser } from '@app/bot/apps/user/parsers/user.parser';
+import { SimpleUser } from '@app/bot/apps/user/parsers/user.types';
 
 @Injectable()
 export class ParseUserPipe implements PipeTransform {
