+++ apps/bot/src/apps/user/user.providers.ts
@@ -1,4 +1,4 @@
-import { UserClient } from './user.client';
-import { UserParser } from './user.parser';
+import { UserClient } from './clients/user.client';
+import { UserParser } from './parsers/user.parser';
 
 export const BOT_USER_PROVIDERS = [UserClient, UserParser];
