+++ apps/bot/src/events/discord-member-joined/discord-member-joined.handler.ts
@@ -1,8 +1,8 @@
-import { GuildGuard } from '@app/bot/guards/guilds/guild.guard';
+import { GuildGuard } from '@app/bot/apps/guild/guards/guild.guard';
 import { Injectable, Logger, UseGuards } from '@nestjs/common';
 import { Context, On } from 'necord';
-import { ParseUserPipe } from '@app/bot/pipes/user/parse-user.pipe';
-import { SimpleUser } from '@app/bot/apps/user/user.types';
+import { ParseUserPipe } from '@app/bot/apps/user/pipes/parse-user.pipe';
+import { SimpleUser } from '@app/bot/apps/user/parsers/user.types';
 
 @UseGuards(GuildGuard)
 @Injectable()
