+++ apps/bot/src/events/offer-message-created/offer-message-created.handler.ts
@@ -1,12 +1,12 @@
 import { Injectable, Logger, UseGuards } from '@nestjs/common';
 import { Context, On } from 'necord';
-import { GuildGuard } from '@app/bot/guards/guilds/guild.guard';
-import { MarketChannelGuard } from '@app/bot/guards/channels/market-channel.guard';
+import { GuildGuard } from '@app/bot/apps/guild/guards/guild.guard';
+import { MarketChannelGuard } from '@app/bot/apps/deal/guards/market-channel.guard';
 import { Type } from '@app/bot/decorators/type.decorator';
-import { ParseUserWithMessagePipe } from '@app/bot/pipes/user/parse-user-with-message.pipe';
-import { ParseCreateOfferInputWithUploadUserImageInputListPipe } from '@app/bot/pipes/offer/parse-create-offer-input-with-upload-user-image-input-list.pipe';
-import { CreateOfferInputWithUploadUserImageInputList } from '@app/bot/apps/offer/offer.types';
-import { OfferClient } from '@app/bot/apps/offer/offer.client';
+import { ParseUserWithMessagePipe } from '@app/bot/apps/user/pipes/parse-user-with-message.pipe';
+import { ParseCreateOfferInputWithUploadUserImageInputListPipe } from '@app/bot/apps/offer/pipes/parse-create-offer-input-with-upload-user-image-input-list.pipe';
+import { CreateOfferInputWithUploadUserImageInputList } from '@app/bot/apps/offer/parsers/offer.types';
+import { OfferClient } from '@app/bot/apps/offer/clients/offer.client';
 
 @UseGuards(GuildGuard, MarketChannelGuard)
 @Type('wts')
